+++ cosmeticFilter.h
@@ -42,11 +42,11 @@ public:
     if (buffer) {
       memcpy(buffer, data, strlen(data) + 1);
     }
-    return strlen(data) + 1;
+    return static_cast<uint32_t>(strlen(data)) + 1;
   }
 
   uint32_t deserialize(char *buffer, uint32_t bufferSize) {
-    int len = strlen(buffer);
+    int len = static_cast<int>(strlen(buffer));
     data = new char[len + 1];
     memcpy(data, buffer, len + 1);
     return len + 1;
@@ -78,7 +78,7 @@ private:
       while (hashItem) {
         CosmeticFilter *cosmeticFilter = hashItem->hashItemStorage;
         // [cosmeticFilter],[space]
-        int cosmeticFilterLen = strlen(cosmeticFilter->data);
+        int cosmeticFilterLen = static_cast<int>(strlen(cosmeticFilter->data));
         if (buffer) {
           memcpy(buffer + len, cosmeticFilter->data, cosmeticFilterLen);
         }
